+++ src/utils/decorators/input-label/__spec__.js
@@ -162,7 +162,7 @@ describe('InputLabel', () => {
     describe('when help is enabled', () => {
       it('adds the relevant class', () => {
         let ExtendedClassOne = InputLabel(HelpClass);
-        let instance = TestUtils.renderIntoDocument(<ExtendedClassOne fieldHelp={ true } />);
+        let instance = TestUtils.renderIntoDocument(<ExtendedClassOne labelHelp={ true } />);
         expect(instance.labelClasses).toEqual('common-input__label common-input__label--help');
       });
     });